requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
firebase-admin==6.5.0
google-generativeai==0.7.2
wikipedia-api==0.6.0
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.7.1
